libjpeg-turbo: avoid CPU feature detection at runtime
The current code was parsing /proc/cpuinfo at runtime,
to find out whether the target CPU was Cortex-A53 to
control the usage of "FASTTBL" which is a regression
on Cortex-A53.

In Android that is unnecessary since we have the
information about the Target CPU provided at compile
time via the Android build system.

This patch introduces the flag "AVOID_CPUINFO_CHECK",
passes the flag for arm64 targets by default and
replaces the runtime detection with compile time detection.
The flag "TARGET_IS_CORTEX_A53" which gets passed based
on "TARGET_CPU_VARIANT" is used to run the compile time detection
on whether to allow using the "FASTTBL" instruction or not.
Detection for Cortex-A57 was removed since there is no
matching supported CPU target in arm64.
Feature detection code for Cavium ThunderX was completely
removed since Android is not supported on that target.

Change-Id: I4699248d6cf668ed2585f66f5db6a3eadc40147f
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>